rn: Rename the user-facing component files
Having two files with the same name is not good, so put a different prefix on these. Also, call the router what it is.
