+++ packages/edge-login-ui-rn/src/index.js
@@ -2,11 +2,11 @@
 
 import './util/androidFetch.js'
 
-export * from './components/publicApi/ChangePasswordScreen.js'
-export * from './components/publicApi/ChangePinScreen.js'
+export * from './components/publicApi/PublicChangePasswordScreen.js'
+export * from './components/publicApi/PublicChangePinScreen.js'
 export * from './components/publicApi/ChooseTestAppScreen.js'
-export * from './components/publicApi/LoginScreen.js'
-export * from './components/publicApi/PasswordRecoveryScreen.js'
+export * from './components/publicApi/PublicLoginScreen.js'
+export * from './components/publicApi/PublicChangeRecoveryScreen.js'
 
 export {
   isTouchEnabled,
